// src/app/modules/user/user.model.ts
import mongoose, { model } from "mongoose";
import { IAuthProvider, IsActive, IUser, Role } from "./user.interface";
import { boolean } from "zod";
const { Schema } = mongoose;

const authProviderSchema = new Schema<IAuthProvider>(
  {
    provider: { type: String, required: true },
    providerId: { type: String, required: true },
  },
  {
    _id: false,
    versionKey: false,
  },
);

const userSchema = new Schema<IUser>(
  {
    name: { type: String, required: true },
    email: { type: String, required: true, unique: true },
    password: { type: String },
    phone: { type: String },
    picture: { type: String },
    address: { type: String },
    isDeleted: { type: boolean, default: false },
    isActive: {
      type: String,
      enum: Object.values(IsActive),
      default: IsActive.ACTIVE,
    },
    isVerified: { type: boolean, default: false },
    role: { type: String, enum: Object.values(Role), default: Role.USER },
    auths: [authProviderSchema],
  },
  {
    timestamps: true,
    versionKey: false, // Disable the __v field in DB
  },
);

const User = model<IUser>("User", userSchema);

export default User;
